+++ src/components/RequesterLinks/index.js
@@ -1,0 +1,2 @@
+export { default as BarcodeLink } from './BarcodeLink';
+export { default as FullNameLink } from './FullNameLink';