+++ src/utils.test.js
@@ -4,7 +4,6 @@ import {
 
 import {
   buildTemplate,
-  createUserHighlightBoxLink,
   duplicateRequest,
   escapeValue,
   getTlrSettings,
@@ -121,22 +120,6 @@ describe('buildTemplate', () => {
     });
 
     expect(v).toEqual('The vorpal blade went snicker-snack!');
-  });
-});
-
-describe('createUserHighlightBoxLink', () => {
-  it('returns a link given values', () => {
-    const text = 't';
-    const id = 'id';
-    const c = createUserHighlightBoxLink(text, id);
-
-    expect(c.props.to).toMatch(`/users/view/${id}`);
-    expect(c.props.children).toMatch(text);
-  });
-
-  it('returns empty string given no values', () => {
-    const text = createUserHighlightBoxLink('', '');
-    expect(text).toMatch('');
   });
 });